unsigned int -> size_t in some appropriate places. Fixes segfaults on x86_64
git-svn-id: https://svn.gnumonks.org/trunk/librfid@1984 e0336214-984f-0b4b-a45f-81c69e1f0ede

+++ src/rfid_reader_cm5121_openct.c
@@ -13,8 +13,8 @@ static int slot = 1;
 /* this is the sole function required by rfid_reader_cm5121.c */
 int 
 PC_to_RDR_Escape(void *handle, 
-		 const unsigned char *tx_buf, unsigned int tx_len,
-		 unsigned char *rx_buf, unsigned int *rx_len)
+		 const unsigned char *tx_buf, size_t tx_len,
+		 unsigned char *rx_buf, size_t *rx_len)
 {
 	int rc;
 	ct_handle *h = (ct_handle *) handle;
